Refactored <Listings> into its own component.
