+++ components/index.js
@@ -1,3 +1,4 @@
 export * from './forms';
 export { default as Header } from './header';
+export { default as Listings } from './listings';
