+++ components/listings/container.js
@@ -2,17 +2,13 @@ import React from 'react';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
 
-import { apply } from 'reputable/dist/redux/action-creators/registry';
-import { getParticipants, getRegistryAddress } from 'reputable/dist/redux/selectors';
+import { getApplications } from 'reputable/dist/redux/selectors';
 
 const mapStateToProps = (state) => ({
-  registryAddress: getRegistryAddress(state),
-  participants: getParticipants(state),
+  applications: getApplications(state),
 });
 
-const mapDispatchToProps = dispatch => bindActionCreators({
-  submitApplication: apply,
-}, dispatch);
+const mapDispatchToProps = dispatch => bindActionCreators({}, dispatch);
 
 const createContainer = (ComposedComponent) => {
   class Container extends React.Component {
